requirements.txt
amqp==2.6.1
asgiref==3.2.10
bcrypt==3.1.7
billiard==3.6.3.0
celery==4.4.7
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cloudinary==1.17.0
cryptography==2.9.2
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==2.0.5
django-admin-sortable2==0.7.7
django-appconf==1.0.4
django-celery-email==3.0.0
django-cloudinary-storage==0.2.3
django-heroku==0.3.1
django-import-export==2.3.0
dotmap==1.3.8
et-xmlfile==1.0.1
Fabric3==1.14.post1
gunicorn==19.7.1
idna==2.10
jdcal==1.4.1
kiwisolver==1.2.0
kombu==4.6.11
MarkupPy==1.14
matplotlib==3.3.0
mock==4.0.2
names==0.3.0
numpy==1.19.1
odfpy==1.4.1
openpyxl==3.0.4
paramiko==2.7.1
Pillow==7.2.0
psycopg2==2.8.4
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.9
vine==1.3.0
whitenoise==4.1.4
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
